refactor: move use-contracts-cache
